requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
Faker==22.5.1
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==7.4.4
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.1.0
